requirements.txt
django==1.6
MySQL-python
djangorestframework
django-bootstrap3
south
django-model-utils